% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terna_energy_balance.R
\name{terna_energy_balance}
\alias{terna_energy_balance}
\title{Get total load from the terna API.}
\usage{
terna_energy_balance(
  date_from,
  date_to,
  type = c("Thermal", "Wind", "Geothermal", "Photovoltaic", "Self-consumption", "Hydro"),
  token = NULL
)
}
\arguments{
\item{date_from}{Date from}

\item{date_to}{Date to}

\item{type}{type.}

\item{token}{token}
}
\description{
Get total load from the terna API.
}
