% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terna_token.R
\name{terna_get_access_token}
\alias{terna_get_access_token}
\title{Get a token to download data.}
\usage{
terna_get_access_token(client_id = NULL, client_secret = NULL)
}
\arguments{
\item{client_id}{Client id. Defaults to NULL.}

\item{client_secret}{Client secret. Defaults to NULL.}
}
\description{
Get a token to download data.
}
