% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terna_installed_capacity.R
\name{terna_installed_capacity}
\alias{terna_installed_capacity}
\title{Get total load from the terna API.}
\usage{
terna_installed_capacity(
  year,
  type = c("Thermal", "Wind", "Geothermal", "Photovoltaic", "Self-consumption", "Hydro"),
  token = NULL
)
}
\arguments{
\item{year}{Year}

\item{type}{type.}

\item{token}{token}
}
\description{
Get total load from the terna API.
}
