% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terna_load.R
\name{terna_load}
\alias{terna_load}
\title{Get total load from the terna API.}
\usage{
terna_load(
  date_from,
  date_to,
  bidding_zone = c("Centre-North", "Centre-South", "North", "Sardinia", "Sicily",
    "South", "Calabria", "Italy"),
  token = NULL
)
}
\arguments{
\item{date_from}{Date from}

\item{date_to}{Date to}

\item{bidding_zone}{Bidding zone.}

\item{token}{token}
}
\description{
Get total load from the terna API.
}
